requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.0.4; python_version >= '3'
cycler==0.10.0
fsspec[http]==2021.7.0; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
git+https://github.com/Yotsuyubi/nn_simulator@5f0c4dbb458da0b8e4222b3b5c4df322c84c228f#egg=nn-simulator
google-auth-oauthlib==0.4.5; python_version >= '3.6'
google-auth==1.35.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
grpcio==1.39.0
idna==3.2; python_version >= '3'
importlib-metadata==4.7.0; python_version < '3.8'
kiwisolver==1.3.1; python_version >= '3.6'
markdown==3.3.4; python_version >= '3.6'
matplotlib==3.4.3
multidict==5.1.0; python_version >= '3.6'
numpy==1.21.2
oauthlib==3.1.1; python_version >= '3.6'
packaging==21.0; python_version >= '3.6'
pillow==8.3.1; python_version >= '3.6'
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydeprecate==0.3.1; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytorch-lightning==1.4.4
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0; python_version >= '3.6'
torch==1.9.0
torchmetrics==0.5.0; python_version >= '3.6'
torchvision==0.10.0
tqdm==4.62.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0; python_version < '3.8'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
wheel==0.37.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.6.3; python_version >= '3.6'
zipp==3.5.0; python_version >= '3.6'
